+++ src/main/java/com/vonage/client/messages/mms/MmsAudioRequest.java
@@ -23,8 +23,6 @@ public final class MmsAudioRequest extends MmsRequest {
 
 	MmsAudioRequest(Builder builder) {
 		super(builder, MessageType.AUDIO);
-		payload = new MessagePayload(builder.url, builder.caption);
-		payload.validateCaptionLength(2000);
 	}
 
 	@JsonProperty("audio")