+++ src/main/java/com/vonage/client/messages/mms/MmsVideoRequest.java
@@ -23,8 +23,6 @@ public final class MmsVideoRequest extends MmsRequest {
 
 	MmsVideoRequest(Builder builder) {
 		super(builder, MessageType.VIDEO);
-		payload = new MessagePayload(builder.url, builder.caption);
-		payload.validateCaptionLength(2000);
 	}
 
 	@JsonProperty("video")