+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 # [8.7.0] - 2024-05-??
 - Added missing supported languages to `TextToSpeechLanguage` enum
+- Added `ttl` field to outbound MMS messages
+- 
 
 # [8.6.0] - 2024-04-18
 - Added Experience Composer to Video API